Remove some `deepcopy`s; ~81ms per cycle
Karger's algorithm is probabilistic, so we can't be sure how many times we'll need to run the contraction process on the input graph before we hit upon a minimal cut. Each run of the process takes ~81 ms, but in 5 runs of `main_a` the number of attempts required was 17, 41, 49, 71, and 307, giving runtimes between 1.4s and 25s.